Fixes #19 with a possible API candidate
There are two halves to this:

Firstly, this seems to addresses issue #19 and the API looks like
it's working and pretty easy to use if you like it.  (Maybe you have
a better idea how to do it, and that's fine if you do.)  Included
in the PR is an example of how to call the API, in a #DEBUG section
with the class CTBWindowsTest.

Secondly, that example in the PR also serves an ulterior motive.
I wanted to have an quick crude example that demonstrates what I
think is a bug in CTB that's been giving me trouble in kOS.  So
I made the small example replicate the same basic usage pattern that
makes kOS trigger the issue, but without all the exteraneous kOS stuff
in the way masking the issue.